requirements.txt
requests==2.26.0
configparser==5.0.2
python-telegram-bot==14.2
